+++ AppVersion.cs
@@ -2,8 +2,8 @@ namespace NOCS
 {
     internal static class AppVersion
     {
-        public const string ReleaseBase = "0.5.18";
+        public const string ReleaseBase = "0.5.23";
         public const string BepInSemVer = ReleaseBase;
-        public const string DisplayVersion = "0.5.18QV";
+        public const string DisplayVersion = "0.5.23QV";
     }
 }